refactor: fetch the similar part of ComposeFloatingWindow.kt and ComposeServiceFloatingWindow.kt into CoreFloatingWindow.kt
